+++ packages/eas-cli/src/commands/update/view.ts
@@ -1,12 +1,10 @@
 import chalk from 'chalk';
-import Table from 'cli-table3';
 
 import EasCommand from '../../commandUtils/EasCommand';
 import { EasJsonOnlyFlag } from '../../commandUtils/flags';
 import { UpdateQuery } from '../../graphql/queries/UpdateQuery';
 import Log from '../../log';
-import { UPDATE_COLUMNS, getUpdateGroupDescriptions } from '../../update/utils';
-import formatFields from '../../utils/formatFields';
+import { formatUpdateGroup, getUpdateGroupDescriptions } from '../../update/utils';
 import { enableJsonOutput, printJsonOnlyOutput } from '../../utils/json';
 
 export default class UpdateView extends EasCommand {
@@ -48,20 +46,9 @@ export default class UpdateView extends EasCommand {
     if (jsonFlag) {
       printJsonOnlyOutput(updateGroupDescription);
     } else {
-      const groupTable = new Table({
-        head: [...UPDATE_COLUMNS],
-        wordWrap: true,
-      });
-
       Log.log(chalk.bold('Update group:'));
-      Log.log(formatFields([{ label: 'ID', value: updateGroupDescription.group }]));
-      groupTable.push([
-        updateGroupDescription.message,
-        updateGroupDescription.runtimeVersion,
-        updateGroupDescription.group,
-        updateGroupDescription.platforms,
-      ]);
-      Log.log(groupTable.toString());
+
+      Log.log(formatUpdateGroup(updateGroupDescription));
     }
   }
 }