+++ packages/eas-cli/src/project/publish.ts
@@ -14,7 +14,7 @@ import { PresignedPost } from '../graphql/mutations/UploadSessionMutation';
 import { PublishQuery } from '../graphql/queries/PublishQuery';
 import Log from '../log';
 import { uploadWithPresignedPostWithRetryAsync } from '../uploads';
-import { expoCommandAsync } from '../utils/expoCli';
+import { expoCommandAsync, shouldUseVersionedExpoCLI } from '../utils/expoCli';
 import chunk from '../utils/expodash/chunk';
 import uniqBy from '../utils/expodash/uniqBy';
 
@@ -160,14 +160,19 @@ export async function buildBundlesAsync({
     throw new Error('Could not locate package.json');
   }
 
-  await expoCommandAsync(projectDir, [
-    'export',
-    '--output-dir',
-    inputDir,
-    '--experimental-bundle',
-    '--non-interactive',
-    '--dump-sourcemap',
-  ]);
+  if (shouldUseVersionedExpoCLI(projectDir)) {
+    await expoCommandAsync(projectDir, ['export', '--output-dir', inputDir, '--dump-sourcemap']);
+  } else {
+    // Legacy global Expo CLI
+    await expoCommandAsync(projectDir, [
+      'export',
+      '--output-dir',
+      inputDir,
+      '--experimental-bundle',
+      '--non-interactive',
+      '--dump-sourcemap',
+    ]);
+  }
 }
 
 export async function resolveInputDirectoryAsync(